Creates types folder and moved files
