+++ client/src/context/auth-context/AuthContext.tsx
@@ -1,7 +1,7 @@
 // The AuthContext provides a state, which indicates whether the user is authenticated or not
 
 import React, { createContext, useReducer } from "react";
-import { Auth, AuthContextType, authReducerAction, authReducerActionType } from "./globalTypes";
+import { Auth, AuthContextType, authReducerAction, authReducerActionType } from "../../types/globalTypes";
 
 /**
  * The initial state of the auth object
