+++ client/src/types/globalTypes.ts
